requirements.txt
web3
pandas
requests
python-decouple 